requirements.txt
Django==4.2.27
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
python-dotenv==1.2.1

paddleocr==3.3.2
paddlepaddle==3.2.0
opencv-contrib-python==4.10.0.84
numpy==2.0.2
pillow==11.3.0
shapely==2.0.7
pyclipper==1.3.0.post6
python-bidi==0.6.7

reportlab==4.4.7

google-genai==1.47.0
google-auth==2.45.0
httpx==0.28.1
tenacity==9.1.2
requests==2.32.5

aiohttp==3.13.2
async-timeout==4.0.3
attrs==25.4.0
frozenlist==1.8.0
multidict==6.7.0
yarl==1.22.0

typing_extensions==4.15.0
six==1.17.0
packaging==25.0
tqdm==4.67.1
